ai-agents/now-ai-agent/pyproject.toml
[project]
name = "now-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[all,http-server]>=0.3.10",
    "fastapi>=0.120.0",
    "langchain>=1.0.2",
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.1",
    "pydantic>=2.12.3",
    "uvicorn>=0.38.0",
]
